refactor(web2): rename components
